Added map() to update <Post/>; Added array of objects in state hook
The props from <Post/> have been moved into an array of two objects, wrapped in a State hook (used to track changes). Then a posts.map function has been added that loops through the array of objects when it changes and renders the values as props. That way the app does that same thing as it did in the previous version, but more efficiently and with less code.
Some comments have also been added