Efficiency update
Added doc.id do the newly created object in the docs.map function. The id is referenced in the posts.map function and used as a key. That way, when the database updates, only the data that has changed will be loaded or updated  in the app, as opposed to updating/loading everything again.