sign in functionality fully working
on line 38 in Header.js added a listener that checks for changes in authentication (line 39) (log in, log out, etc.)
Then it checks if the user exists and if they do, their username is not updated (lines 45,45) but if they dont, their username is updated. (line 50)
setUser is always null. if the user logs in, setUser is set to authUser, which comes from the database. (line 43)
if the user logs out, setUser is reverted to null (line 55)