Sign in, Sign up, Log out functionality finished
Added signIn function (line 24). Just like the signUp function (line 10), it uses the firebase tools auth.signInWithEmailAndPassword(email, password). The email and password here are passed from the states (line 45 - 52). On top of that if there is an error, the function will pass it into an alert and show it on screen. An example of an error is that the email address is already taken by another user.
Added new state on line 48. Its initial value is false and its only changed to true when the button on line 78 is pressed. This state opens the new sign in modal (line 135)
Added buttons that Log out the users (lines 75,123,164). They use the firebase function auth.signOut().