Added progress tracking functionality
on line 58 started writing the handleUpload function
first part is explained in comments
on line 63 added the progress tracking functionality.
It listens for changes in the state and takes snapshots (constantly updates) and gives a number on each update that is based on the bytes transfered devided by the total amount of bytes.